Remove simplified parametrization

+++ tests/test_system_setup_starter.py
@@ -8,18 +8,19 @@ parameters_json = {
     "path_to_module": "examples/household_1_advanced_hp_diesel_car.py",
     "function_in_module": "household_1_advanced_hp_diesel_car",
     "config_class_name": "HouseholdAdvancedHPDieselCarConfig",
-    "simple_parameters": False,
-    "cost_parameters": {"electricity_price": 0.2, "gas_price": 0.1},
     "building_type": "some_building",
-    "number_of_people": 4,
-    "heat_pump_power": 12300,
+    # "some_subconf": {"some_attribute": "some_building"},  # Raises AttributeError
+    "hp_config": {
+        "set_thermal_output_power_in_watt": 12000,
+    },
 }
 
 
 def test_system_setup_starter():
     # Run simulation from config_json
-    result_directory = Path("test_system_setup_starter/result")
-    shutil.rmtree(result_directory)
+    result_directory: Path = Path("test_system_setup_starter/result")
+    if result_directory.is_dir():
+        shutil.rmtree(result_directory)
     result_directory.mkdir(parents=True)
 
     (
@@ -30,7 +31,6 @@ def test_system_setup_starter():
     ) = make_system_setup(
         parameters_json=parameters_json,
         result_directory=result_directory,
-        simplified=True,
     )
 
     main(
